// Projects/Tamin/Eservices-UI/src/app/modules/fac/models/alo-request-det-stackholders.ts
import {AloRequest} from './alo-request';
import {Branch} from './branch';

export class AloRequestDetStackholders {
  detStackholdersId: number;
  firstName: string;
  nationality: string;
  lastName: string;
  position: string;
  movazaf: string;
  nationCode: string;
  character: string;
  positionDesk: string;
  birthDate: Date;
  startDate: Date;
  endDate: Date;
  requests: AloRequest;
  sigOwner: string;
  positiont: string;
 }